+++ JS/ui/elements.js
@@ -1,0 +1,19 @@
+export {createBtn, buyBtn, checkBtn, resetBtn, closeBtn, lotto, money, count, modal, div3, div4, div5, div5b, div6, total}
+
+const createBtn = document.querySelector("#createNum");
+const buyBtn = document.querySelector("#buyLotto");
+const checkBtn = document.querySelector("#resultButton");
+const resetBtn = document.querySelector("#reset");
+const closeBtn = document.querySelector("#modalClose");
+
+const lotto = document.querySelector(".lotto");
+const money = document.querySelector("#result");
+const count =  document.querySelector("#countId");
+
+const modal = document.querySelector("#resultModal");
+const div3 = document.querySelector("#num3");
+const div4 = document.querySelector("#num4");
+const div5 = document.querySelector("#num5");
+const div5b = document.querySelector("#num5b");
+const div6 = document.querySelector("#num6");
+const total = document.querySelector("#total");
